refactor: Coach에서 이름은 따로 상태가 관리되기에CoachName 분리
