+++ src/main/java/menu/domain/Coach.java
@@ -5,24 +5,16 @@ import java.util.Collections;
 import java.util.List;
 
 public class Coach {
-    private final String name;
+    private final CoachName name;
     private final List<String> cannotEatMenus;
     private final List<String> eatMenus = new ArrayList<>();
 
-    private final String NAME_SIZE_INVALID = "[ERROR] 코치의 이름은 2글자에서 4글자만 가능합니다.";
     private final String CANNOT_EAT_MENU_SIZE_INVALID = "[ERROR] 코치가 못 먹는 음식은 2개 이상이면 안됩니다.";
 
-    public Coach(String name, List<String> cannotEatMenus) {
-        validateName(name);
+    public Coach(CoachName name, List<String> cannotEatMenus) {
         validateCannotEatMenus(cannotEatMenus);
         this.name = name;
         this.cannotEatMenus = cannotEatMenus;
-    }
-
-    private void validateName(String name) {
-        if (name.length() < 2 || name.length() > 5) {
-            throw new IllegalArgumentException(NAME_SIZE_INVALID);
-        }
     }
 
     private void validateCannotEatMenus(List<String> menus) {
@@ -53,6 +45,6 @@ public class Coach {
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 }
